% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iRAP_build_data.R
\name{iRAP_build_data}
\alias{iRAP_build_data}
\title{Build an iRap formatted data file from raw files}
\usage{
iRAP_build_data(
  dates,
  lookups,
  join_vars,
  agespecs = NULL,
  renames,
  keepvars,
  indicator,
  datasource,
  SHA = "main"
)
}
\arguments{
\item{dates}{date(s) of the file(s) you wish to use as source data files. Should be a character vector of dates in the format "yyyymmdd"}

\item{lookups}{a character vector of the lookups you want to use for creating new variables}

\item{join_vars}{a character vector of variable names that you want to use to join to the lookup tables}

\item{agespecs}{a list containing data frame(s) each of which specifies age bands to use}

\item{renames}{a data frame showing the old and new names of variables to rename}

\item{keepvars}{a character vector of variables you want to keep from the original dataset}

\item{indicator}{a string showing the name you want to give to the final count variable}

\item{datasource}{a string showing which data source you want to use. The only accepted value currently is 'prison_pop'}
}
\value{
A data frame in iRAP format
}
\description{
This function builds an iRAP formatted data file. 
It combines any number of raw data files into a standard iRAP format
This function is only currently tested with prison population data
}
