% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iRAP_build_table.R
\name{iRAP_build_table}
\alias{iRAP_build_table}
\title{Build an iRAP Table}
\usage{
iRAP_build_table(
  tabledata,
  filtervars = NULL,
  nestedvars = NULL,
  nototalvars = NULL,
  count_indicator,
  mean_indicator = NULL
)
}
\arguments{
\item{tabledata}{An iRAP data file produced by I-RAP::iRAP_build_data()}

\item{filtervars}{A character vector of variables you wish to include in the table}

\item{nestedvars}{A list of vectors of 'nested' variables you wish to include in the table. Nested variables are a set of variables which nest within one another rather than existing independently.}

\item{indicator}{The indicator variable from the table}
}
\value{
A data table containing the final iRAP table
}
\description{
This function take an iRAP data file, produced via the I-RAP::iRAP_build_data() function and produces an iRAP table.
It will provide a breakdown of the data by specified varaibles, including creating 'Total' rows.
}
