% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prison_pop_data.R
\name{prison_pop_data}
\alias{prison_pop_data}
\title{Build a prison population dataset in iRAP format}
\usage{
prison_pop_data(dates)
}
\arguments{
\item{dates}{date(s) of the file(s) you wish to use as source data files. Should be a character vector of dates in the format "yyyymmdd"}
}
\value{
a data frame of prison population data
}
\description{
This function builds an iRAP formatted data file of the prison population
It is a specific form of the iRAP_build_data() function with specified parameters for the prison population data file
}
