% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prison_pop_table.R
\name{prison_pop_table}
\alias{prison_pop_table}
\title{Build a prison population table}
\usage{
prison_pop_table(tabledata, filtervars, nestedvars = NULL)
}
\arguments{
\item{tabledata}{A prison population data file}

\item{filtervars}{A character vector of variables you wish to include in the table}

\item{nestedvars}{A list of vectors of 'nested' variables you wish to include in the table. Nested variables are a set of variables which nest within one another rather than existing independently.}
}
\value{
A data table showing the final table data
}
\description{
This function take a prison population data file and can be used to build publication table files.
It will provide a breakdown of the data by specified varaibles, including creating 'Total' rows.
}
