% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prison_pop_tables.R
\name{prison_pop_tables}
\alias{prison_pop_tables}
\title{Build a set of prison population data tables in iRAP format}
\usage{
prison_pop_tables(dates)
}
\arguments{
\item{dates}{date(s) of the file(s) you wish to use as source data files. Should be a character vector of dates in the format "yyyymmdd"}
}
\value{
a list of data tables containing standard published data
}
\description{
This function builds a set of iRAP formatted data tables of the prison population
It is a specific form of the iRAP_build_table() function with specified parameters for the standard published prison population tables
}
