% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prison_releases_data.R
\name{prison_releases_data}
\alias{prison_releases_data}
\title{Build a prison releases dataset in iRAP format}
\usage{
prison_releases_data(dates, SHA = "main")
}
\arguments{
\item{dates}{date(s) of the file(s) you wish to use as source data files. Should be a character vector of dates in the format "yyyyQq"}
}
\value{
a data frame of prison releases data
}
\description{
This function builds an iRAP formatted data file of prison releases
It is a specific form of the iRAP_build_data() function with specified parameters for the prison releases data file
}
