% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prison_pop_data.R
\name{prison_pop_data}
\alias{prison_pop_data}
\title{Build a prison population data file from raw SAS files}
\usage{
prison_pop_data(dates)
}
\arguments{
\item{dates}{date of the population file(s) you wish to build a complete file from. Should be a character vector of dates in the format "yyyymmdd"}
}
\value{
A data frame of the prison population for each date selected
}
\description{
This function builds a prison population data file. 
It can combine multiple files based on population extracts taken on different dates.
}
